started the cm matrices
We have finished the cm and the forward selection. there are two versions of the forward selection: one w logit regression and one with old regression. I think we will want to do the logit regression but idk if rpsuedo is right. now we should do some transformations and interactions for the regression. We should prob maximize threshold too but idk in reference to what